changed package dir
