+++ neutronics-material-maker/data/magnet_materials.json
